HashMap
HashMap is a part of Java’s collection since Java 1.2. It provides the basic implementation of the Map interface of Java. It stores the data in (Key, Value) pairs.

These are various important hashmap class methods. This post explains: put(), get(), isEmpty() and size()

put(): java.util.HashMap.put() plays role in associating the specified value with the specified key in this map. If the map previously contained a mapping for the key, the old value is replaced.

get(): java.util.HashMap.get()method returns the value to which the specified key is mapped, or null if this map contains no mapping for the key.

isEmpty(): java.util.HashMap.isEmpty() method returns true if the map contains no key-value mappings.

size(): java.util.HashMap.size() returns the number of key-value mappings in this map.